requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
transliterate==1.10.2
urllib3==1.26.9
